pyproject.toml
[tool.poetry]
name = "grammys-simulation"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "grammys_simulation"}]

[tool.poetry.dependencies]
python = "^3.10"
streamlit = "^1.18.1"
pandas = "^1.5.3"
inflect = "^6.0.2"
tqdm = "^4.64.1"
stqdm = "^0.0.5"
altair = "^4.2.2"
pyyaml = "^6.0"


[tool.poetry.group.dev.dependencies]
watchdog = "^2.2.1"
jupyterlab = "^3.6.1"
ipykernel = "^6.21.2"
matplotlib = "^3.7.0"
seaborn = "^0.12.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
